requirements.txt
Package             Version
------------------- ------------
blinker             1.7.0
cachelib            0.9.0
certifi             2023.7.22
charset-normalizer  3.3.2
click               8.1.7
Deprecated          1.2.14
Flask               3.0.0
Flask-Caching       2.1.0
Flask-Limiter       3.5.0
Flask-SQLAlchemy    3.1.1
greenlet            3.0.1
idna                3.4
importlib-resources 6.1.1
itsdangerous        2.1.2
Jinja2              3.1.2
limits              3.7.0
markdown-it-py      3.0.0
MarkupSafe          2.1.3
mdurl               0.1.2
multipledispatch    1.0.0
numpy               1.26.2
ordered-set         4.1.0
packaging           23.2
pip                 23.3.1
psycopg2-binary     2.9.9
Pygments            2.17.2
python-dateutil     2.8.2
python-dotenv       1.0.0
pytz                2023.3.post1
requests            2.31.0
rich                13.7.0
setuptools          65.5.0
six                 1.16.0
SQLAlchemy          2.0.23
typing_extensions   4.8.0
tzdata              2023.3
urllib3             2.1.0
Werkzeug            3.0.1
wrapt               1.16.0
